refactor: move store utils to shared folder
