+++ src/account/components/OptionalAuth.js
@@ -15,10 +15,12 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 import React, { useCallback } from 'react';
+
 import _ from 'lodash/fp';
 import { useSelector } from 'react-redux';
 import { fetchUser } from 'terrasoApi/shared/account/accountSlice';
-import { useFetchData } from 'terrasoApi/utils';
+import { useFetchData } from 'terrasoApi/shared/store/utils';
+
 import PageLoader from 'layout/PageLoader';
 
 const OptionalAuth = ({ children }) => {
