+++ src/sharedData/sharedDataSlice.js
@@ -16,7 +16,8 @@
  */
 import { createSlice } from '@reduxjs/toolkit';
 import _ from 'lodash/fp';
-import { createAsyncThunk } from 'terrasoApi/utils';
+import { createAsyncThunk } from 'terrasoApi/shared/store/utils';
+
 import * as sharedDataService from 'sharedData/sharedDataService';
 
 export const UPLOAD_STATUS_UPLOADING = 'uploading';
