+++ src/storyMap/storyMapSlice.js
@@ -16,7 +16,8 @@
  */
 import { createSlice } from '@reduxjs/toolkit';
 import _ from 'lodash/fp';
-import { createAsyncThunk } from 'terrasoApi/utils';
+import { createAsyncThunk } from 'terrasoApi/shared/store/utils';
+
 import * as storyMapService from 'storyMap/storyMapService';
 
 const initialState = {
